refactor(anthropic): split provider into per-concern peer files
internal/providers/anthropic/anthropic.go was 1582 lines mixing
provider plumbing, chat completion, chat streaming, responses API
(non-streaming + streaming), and the full batch lifecycle. Split
into peer files in the same package:

  chat.go (85)         ChatCompletion, convertFromAnthropicResponse
  chat_stream.go (362) StreamChatCompletion, streamConverter
  responses.go (382)   Responses, StreamResponses,
                       responsesStreamConverter, responses-usage helpers
  batch.go (366)       Create/Get/List/Cancel + result parsing,
                       parseOptionalUnix, mapAnthropicBatchResponse

anthropic.go (448) keeps Provider struct + lifecycle helpers,
ListModels, content extractors, shared stream/usage helpers, and
the no-op Embeddings stub. Pure relocation; tests pass.
