+++ internal/providers/anthropic/anthropic.go
@@ -2,22 +2,16 @@
 package anthropic
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"io"
 	"log/slog"
 	"maps"
 	"net/http"
-	"net/url"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
-
-	"github.com/google/uuid"
 
 	"gomodel/internal/core"
 	"gomodel/internal/llmclient"
@@ -223,536 +217,6 @@ func normalizeEffort(effort string) string {
 	}
 }
 
-// convertFromAnthropicResponse converts Anthropic response to core.ChatResponse
-func convertFromAnthropicResponse(resp *anthropicResponse) *core.ChatResponse {
-	content := extractTextContent(resp.Content)
-	thinking := extractThinkingContent(resp.Content)
-	toolCalls := extractToolCalls(resp.Content)
-
-	finishReason := normalizeAnthropicStopReason(resp.StopReason)
-	if finishReason == "" {
-		finishReason = "stop"
-	}
-
-	usage := core.Usage{
-		PromptTokens:     resp.Usage.InputTokens,
-		CompletionTokens: resp.Usage.OutputTokens,
-		TotalTokens:      resp.Usage.InputTokens + resp.Usage.OutputTokens,
-	}
-
-	rawUsage := buildAnthropicRawUsage(resp.Usage)
-	if len(rawUsage) > 0 {
-		usage.RawUsage = rawUsage
-	}
-
-	msg := core.ResponseMessage{
-		Role:      "assistant",
-		Content:   content,
-		ToolCalls: toolCalls,
-	}
-
-	// Surface thinking content as reasoning_content (OpenAI-compatible format).
-	if thinking != "" {
-		raw, err := json.Marshal(thinking)
-		if err == nil {
-			msg.ExtraFields = core.UnknownJSONFieldsFromMap(map[string]json.RawMessage{
-				"reasoning_content": raw,
-			})
-		}
-	}
-
-	return &core.ChatResponse{
-		ID:      resp.ID,
-		Object:  "chat.completion",
-		Model:   resp.Model,
-		Created: time.Now().Unix(),
-		Choices: []core.Choice{
-			{
-				Index:        0,
-				Message:      msg,
-				FinishReason: finishReason,
-			},
-		},
-		Usage: usage,
-	}
-}
-
-// ChatCompletion sends a chat completion request to Anthropic
-func (p *Provider) ChatCompletion(ctx context.Context, req *core.ChatRequest) (*core.ChatResponse, error) {
-	anthropicReq, err := convertToAnthropicRequest(req)
-	if err != nil {
-		return nil, err
-	}
-
-	var anthropicResp anthropicResponse
-	err = p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/messages",
-		Body:     anthropicReq,
-	}, &anthropicResp)
-	if err != nil {
-		return nil, err
-	}
-
-	return convertFromAnthropicResponse(&anthropicResp), nil
-}
-
-// StreamChatCompletion returns a raw response body for streaming (caller must close)
-func (p *Provider) StreamChatCompletion(ctx context.Context, req *core.ChatRequest) (io.ReadCloser, error) {
-	anthropicReq, err := convertToAnthropicRequest(req)
-	if err != nil {
-		return nil, err
-	}
-	anthropicReq.Stream = true
-
-	stream, err := p.client.DoStream(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/messages",
-		Body:     anthropicReq,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// Return a reader that converts Anthropic SSE format to OpenAI format
-	return newStreamConverter(stream, req.Model), nil
-}
-
-// streamConverter wraps an Anthropic stream and converts it to OpenAI format
-type streamConverter struct {
-	reader            *bufio.Reader
-	body              io.ReadCloser
-	model             string
-	msgID             string
-	nextToolCallIndex int
-	toolCalls         map[int]*streamToolCallState
-	thinkingBlocks    map[int]bool // tracks which content block indices are thinking blocks
-	usage             anthropicUsage
-	hasUsage          bool
-	buffer            streaming.StreamBuffer
-	closed            bool
-	emittedToolCalls  bool
-}
-
-type streamToolCallState struct {
-	ID                string
-	Name              string
-	Arguments         strings.Builder
-	Index             int
-	Started           bool
-	PlaceholderObject bool
-}
-
-func newStreamConverter(body io.ReadCloser, model string) *streamConverter {
-	return &streamConverter{
-		reader:         bufio.NewReader(body),
-		body:           body,
-		model:          model,
-		toolCalls:      make(map[int]*streamToolCallState),
-		thinkingBlocks: make(map[int]bool),
-		buffer:         streaming.NewStreamBuffer(1024),
-	}
-}
-
-func malformedAnthropicStreamError(err error) error {
-	return core.NewProviderError("anthropic", http.StatusBadGateway, "failed to decode anthropic stream event: "+err.Error(), err)
-}
-
-func consumeAnthropicSSELine(p []byte, line []byte, body io.ReadCloser, buffer *streaming.StreamBuffer, convert func(*anthropicStreamEvent) string) (n int, handled bool, err error) {
-	line = bytes.TrimSpace(line)
-	if len(line) == 0 || bytes.HasPrefix(line, []byte("event:")) {
-		return 0, false, nil
-	}
-	if !bytes.HasPrefix(line, []byte("data:")) {
-		return 0, false, nil
-	}
-
-	data := bytes.TrimSpace(bytes.TrimPrefix(line, []byte("data:")))
-
-	var event anthropicStreamEvent
-	if err := json.Unmarshal(data, &event); err != nil {
-		_ = body.Close() //nolint:errcheck
-		return 0, false, malformedAnthropicStreamError(err)
-	}
-
-	chunk := convert(&event)
-	if chunk == "" {
-		return 0, false, nil
-	}
-
-	buffer.AppendString(chunk)
-	return buffer.Read(p), true, nil
-}
-
-func mergeAnthropicUsage(dst *anthropicUsage, src *anthropicUsage) bool {
-	if dst == nil || src == nil {
-		return false
-	}
-
-	merged := false
-	if src.InputTokens != 0 {
-		dst.InputTokens = src.InputTokens
-		merged = true
-	}
-	if src.OutputTokens != 0 {
-		dst.OutputTokens = src.OutputTokens
-		merged = true
-	}
-	if src.CacheCreationInputTokens != 0 {
-		dst.CacheCreationInputTokens = src.CacheCreationInputTokens
-		merged = true
-	}
-	if src.CacheReadInputTokens != 0 {
-		dst.CacheReadInputTokens = src.CacheReadInputTokens
-		merged = true
-	}
-
-	return merged
-}
-
-func anthropicChatUsagePayload(usage *anthropicUsage) map[string]any {
-	if usage == nil {
-		return nil
-	}
-
-	payload := map[string]any{
-		"prompt_tokens":     usage.InputTokens,
-		"completion_tokens": usage.OutputTokens,
-		"total_tokens":      usage.InputTokens + usage.OutputTokens,
-	}
-	if usage.CacheReadInputTokens > 0 {
-		payload["cache_read_input_tokens"] = usage.CacheReadInputTokens
-	}
-	if usage.CacheCreationInputTokens > 0 {
-		payload["cache_creation_input_tokens"] = usage.CacheCreationInputTokens
-	}
-	return payload
-}
-
-func anthropicResponsesUsagePayload(usage *anthropicUsage) map[string]any {
-	if usage == nil {
-		return nil
-	}
-
-	payload := map[string]any{
-		"input_tokens":  usage.InputTokens,
-		"output_tokens": usage.OutputTokens,
-		"total_tokens":  usage.InputTokens + usage.OutputTokens,
-	}
-	if usage.CacheReadInputTokens > 0 {
-		payload["cache_read_input_tokens"] = usage.CacheReadInputTokens
-	}
-	if usage.CacheCreationInputTokens > 0 {
-		payload["cache_creation_input_tokens"] = usage.CacheCreationInputTokens
-	}
-	return payload
-}
-
-func (sc *streamConverter) Read(p []byte) (n int, err error) {
-	// If we have buffered data, return it first
-	if sc.buffer.Len() > 0 {
-		return sc.buffer.Read(p), nil
-	}
-
-	if sc.closed {
-		sc.releaseBuffer()
-		return 0, io.EOF
-	}
-
-	// Read the next SSE event from Anthropic
-	for {
-		line, err := sc.reader.ReadBytes('\n')
-		if err != nil {
-			if err == io.EOF {
-				// Send final [DONE] message
-				sc.buffer.AppendString("data: [DONE]\n\n")
-				n = sc.buffer.Read(p)
-				sc.closed = true
-				_ = sc.body.Close() //nolint:errcheck
-				return n, nil
-			}
-			return 0, err
-		}
-
-		n, handled, err := consumeAnthropicSSELine(p, line, sc.body, &sc.buffer, sc.convertEvent)
-		if err != nil {
-			sc.closed = true
-			sc.releaseBuffer()
-			return 0, err
-		}
-		if handled {
-			if n == 0 {
-				continue
-			}
-			return n, nil
-		}
-	}
-}
-
-func (sc *streamConverter) Close() error {
-	if sc.closed {
-		sc.releaseBuffer()
-		return nil
-	}
-	sc.closed = true
-	sc.releaseBuffer()
-	return sc.body.Close()
-}
-
-func (sc *streamConverter) releaseBuffer() {
-	sc.buffer.Release()
-}
-
-func (sc *streamConverter) mapStreamStopReason(reason string) string {
-	// Preserve raw "tool_use" when the upstream stream never produced any
-	// tool call deltas. This avoids claiming OpenAI-style tool calls for a
-	// malformed or partial Anthropic stream.
-	if reason == "tool_use" && !sc.emittedToolCalls {
-		return reason
-	}
-	return normalizeAnthropicStopReason(reason)
-}
-
-func extractInitialToolArguments(input json.RawMessage) string {
-	if len(input) == 0 {
-		return ""
-	}
-
-	trimmed := strings.TrimSpace(string(input))
-	if trimmed == "" || trimmed == "null" {
-		return ""
-	}
-
-	var parsed any
-	if err := json.Unmarshal(input, &parsed); err != nil {
-		return trimmed
-	}
-
-	canonical, err := json.Marshal(parsed)
-	if err != nil {
-		return trimmed
-	}
-
-	return string(canonical)
-}
-
-func normalizeAnthropicStopReason(stopReason string) string {
-	switch stopReason {
-	case "tool_use":
-		return "tool_calls"
-	case "end_turn", "stop_sequence":
-		return "stop"
-	case "max_tokens", "model_context_window_exceeded":
-		return "length"
-	default:
-		return stopReason
-	}
-}
-
-func (sc *streamConverter) formatChatChunk(delta map[string]any, finishReason any, usage *anthropicUsage) string {
-	chunk := map[string]any{
-		"id":       sc.msgID,
-		"object":   "chat.completion.chunk",
-		"created":  time.Now().Unix(),
-		"model":    sc.model,
-		"provider": "anthropic",
-		"choices": []map[string]any{
-			{
-				"index":         0,
-				"delta":         delta,
-				"finish_reason": finishReason,
-			},
-		},
-	}
-	if usage != nil {
-		chunk["usage"] = anthropicChatUsagePayload(usage)
-	}
-
-	jsonData, err := json.Marshal(chunk)
-	if err != nil {
-		slog.Error("failed to marshal chat completion chunk", "error", err, "msg_id", sc.msgID)
-		return ""
-	}
-
-	return fmt.Sprintf("data: %s\n\n", jsonData)
-}
-
-func (sc *streamConverter) convertEvent(event *anthropicStreamEvent) string {
-	switch event.Type {
-	case "message_start":
-		role := ""
-		if event.Message != nil {
-			sc.msgID = event.Message.ID
-			if mergeAnthropicUsage(&sc.usage, &event.Message.Usage) {
-				sc.hasUsage = true
-			}
-			role = strings.TrimSpace(event.Message.Role)
-		}
-		if mergeAnthropicUsage(&sc.usage, event.Usage) {
-			sc.hasUsage = true
-		}
-		if event.Message != nil {
-			if role == "" {
-				role = "assistant"
-			}
-			return sc.formatChatChunk(map[string]any{
-				"role": role,
-			}, nil, nil)
-		}
-		return ""
-
-	case "content_block_start":
-		if event.ContentBlock != nil && event.ContentBlock.Type == "thinking" {
-			sc.thinkingBlocks[event.Index] = true
-			return ""
-		}
-		if event.ContentBlock != nil && event.ContentBlock.Type == "tool_use" {
-			state := &streamToolCallState{
-				ID:    event.ContentBlock.ID,
-				Name:  event.ContentBlock.Name,
-				Index: sc.nextToolCallIndex,
-			}
-			sc.nextToolCallIndex++
-
-			initialArguments := extractInitialToolArguments(event.ContentBlock.Input)
-			state.PlaceholderObject = initialArguments == "{}"
-			if state.PlaceholderObject {
-				sc.toolCalls[event.Index] = state
-				return ""
-			}
-			if initialArguments != "" {
-				_, _ = state.Arguments.WriteString(initialArguments)
-			}
-			state.Started = true
-			sc.toolCalls[event.Index] = state
-			sc.emittedToolCalls = true
-
-			return sc.formatChatChunk(map[string]any{
-				"tool_calls": []map[string]any{
-					{
-						"index": state.Index,
-						"id":    state.ID,
-						"type":  "function",
-						"function": map[string]any{
-							"name":      state.Name,
-							"arguments": initialArguments,
-						},
-					},
-				},
-			}, nil, nil)
-		}
-		return ""
-
-	case "content_block_delta":
-		if event.Delta == nil {
-			return ""
-		}
-
-		switch event.Delta.Type {
-		case "thinking_delta":
-			if sc.thinkingBlocks[event.Index] && event.Delta.Thinking != "" {
-				return sc.formatChatChunk(map[string]any{
-					"reasoning_content": event.Delta.Thinking,
-				}, nil, nil)
-			}
-		case "signature_delta":
-			// Signature deltas are internal to Anthropic's thinking protocol;
-			// no OpenAI-compatible equivalent to emit.
-			return ""
-		case "text_delta":
-			if event.Delta.Text != "" {
-				return sc.formatChatChunk(map[string]any{
-					"content": event.Delta.Text,
-				}, nil, nil)
-			}
-		case "input_json_delta":
-			if event.Delta.PartialJSON == "" {
-				return ""
-			}
-			state := sc.toolCalls[event.Index]
-			if state == nil {
-				return ""
-			}
-			if state.PlaceholderObject {
-				state.Arguments = strings.Builder{}
-				state.PlaceholderObject = false
-			}
-			_, _ = state.Arguments.WriteString(event.Delta.PartialJSON)
-			if !state.Started {
-				state.Started = true
-				sc.emittedToolCalls = true
-				return sc.formatChatChunk(map[string]any{
-					"tool_calls": []map[string]any{
-						{
-							"index": state.Index,
-							"id":    state.ID,
-							"type":  "function",
-							"function": map[string]any{
-								"name":      state.Name,
-								"arguments": event.Delta.PartialJSON,
-							},
-						},
-					},
-				}, nil, nil)
-			}
-			sc.emittedToolCalls = true
-			return sc.formatChatChunk(map[string]any{
-				"tool_calls": []map[string]any{
-					{
-						"index": state.Index,
-						"function": map[string]any{
-							"arguments": event.Delta.PartialJSON,
-						},
-					},
-				},
-			}, nil, nil)
-		}
-
-	case "content_block_stop":
-		state := sc.toolCalls[event.Index]
-		if state != nil && !state.Started && state.PlaceholderObject {
-			state.Started = true
-			sc.emittedToolCalls = true
-			return sc.formatChatChunk(map[string]any{
-				"tool_calls": []map[string]any{
-					{
-						"index": state.Index,
-						"id":    state.ID,
-						"type":  "function",
-						"function": map[string]any{
-							"name":      state.Name,
-							"arguments": "{}",
-						},
-					},
-				},
-			}, nil, nil)
-		}
-		return ""
-
-	case "message_delta":
-		if mergeAnthropicUsage(&sc.usage, event.Usage) {
-			sc.hasUsage = true
-		}
-		// Emit chunk if we have stop_reason or usage data
-		if (event.Delta != nil && event.Delta.StopReason != "") || event.Usage != nil {
-			var finishReason any
-			if event.Delta != nil && event.Delta.StopReason != "" {
-				finishReason = sc.mapStreamStopReason(event.Delta.StopReason)
-			}
-			var usage *anthropicUsage
-			if sc.hasUsage {
-				usage = &sc.usage
-			}
-			return sc.formatChatChunk(map[string]any{}, finishReason, usage)
-		}
-
-	case "message_stop":
-		return ""
-	}
-
-	return ""
-}
-
 // ListModels retrieves the list of available models from Anthropic's /v1/models endpoint
 func (p *Provider) ListModels(ctx context.Context) (*core.ModelsResponse, error) {
 	var anthropicResp anthropicModelsResponse
@@ -870,32 +334,6 @@ func extractToolCalls(blocks []anthropicContent) []core.ToolCall {
 	return out
 }
 
-// convertAnthropicResponseToResponses converts an Anthropic response to ResponsesResponse
-func convertAnthropicResponseToResponses(resp *anthropicResponse, model string) *core.ResponsesResponse {
-	content := extractTextContent(resp.Content)
-	toolCalls := extractToolCalls(resp.Content)
-
-	msg := core.Message{
-		Content:   content,
-		ToolCalls: toolCalls,
-	}
-	output := providers.BuildResponsesOutputItems(core.ResponseMessage{
-		Role:      "assistant",
-		Content:   msg.Content,
-		ToolCalls: msg.ToolCalls,
-	})
-
-	return &core.ResponsesResponse{
-		ID:        resp.ID,
-		Object:    "response",
-		CreatedAt: time.Now().Unix(),
-		Model:     model,
-		Status:    "completed",
-		Output:    output,
-		Usage:     buildAnthropicResponsesUsage(resp.Usage),
-	}
-}
-
 // buildAnthropicRawUsage extracts cache fields from anthropicUsage into a RawData map.
 func buildAnthropicRawUsage(u anthropicUsage) map[string]any {
 	raw := make(map[string]any)
@@ -911,672 +349,100 @@ func buildAnthropicRawUsage(u anthropicUsage) map[string]any {
 	return raw
 }
 
-// buildAnthropicResponsesUsage creates a ResponsesUsage from anthropicUsage, including RawUsage.
-func buildAnthropicResponsesUsage(u anthropicUsage) *core.ResponsesUsage {
-	usage := &core.ResponsesUsage{
-		InputTokens:  u.InputTokens,
-		OutputTokens: u.OutputTokens,
-		TotalTokens:  u.InputTokens + u.OutputTokens,
-	}
-	rawUsage := buildAnthropicRawUsage(u)
-	if len(rawUsage) > 0 {
-		usage.RawUsage = rawUsage
-	}
-	return usage
+func malformedAnthropicStreamError(err error) error {
+	return core.NewProviderError("anthropic", http.StatusBadGateway, "failed to decode anthropic stream event: "+err.Error(), err)
 }
 
-// Responses sends a Responses API request to Anthropic (converted to messages format)
-func (p *Provider) Responses(ctx context.Context, req *core.ResponsesRequest) (*core.ResponsesResponse, error) {
-	anthropicReq, err := convertResponsesRequestToAnthropic(req)
+func consumeAnthropicSSELine(p []byte, line []byte, body io.ReadCloser, buffer *streaming.StreamBuffer, convert func(*anthropicStreamEvent) string) (n int, handled bool, err error) {
+	line = bytes.TrimSpace(line)
+	if len(line) == 0 || bytes.HasPrefix(line, []byte("event:")) {
+		return 0, false, nil
+	}
+	if !bytes.HasPrefix(line, []byte("data:")) {
+		return 0, false, nil
+	}
+
+	data := bytes.TrimSpace(bytes.TrimPrefix(line, []byte("data:")))
+
+	var event anthropicStreamEvent
+	if err := json.Unmarshal(data, &event); err != nil {
+		_ = body.Close() //nolint:errcheck
+		return 0, false, malformedAnthropicStreamError(err)
+	}
+
+	chunk := convert(&event)
+	if chunk == "" {
+		return 0, false, nil
+	}
+
+	buffer.AppendString(chunk)
+	return buffer.Read(p), true, nil
+}
+
+func mergeAnthropicUsage(dst *anthropicUsage, src *anthropicUsage) bool {
+	if dst == nil || src == nil {
+		return false
+	}
+
+	merged := false
+	if src.InputTokens != 0 {
+		dst.InputTokens = src.InputTokens
+		merged = true
+	}
+	if src.OutputTokens != 0 {
+		dst.OutputTokens = src.OutputTokens
+		merged = true
+	}
+	if src.CacheCreationInputTokens != 0 {
+		dst.CacheCreationInputTokens = src.CacheCreationInputTokens
+		merged = true
+	}
+	if src.CacheReadInputTokens != 0 {
+		dst.CacheReadInputTokens = src.CacheReadInputTokens
+		merged = true
+	}
+
+	return merged
+}
+
+func extractInitialToolArguments(input json.RawMessage) string {
+	if len(input) == 0 {
+		return ""
+	}
+
+	trimmed := strings.TrimSpace(string(input))
+	if trimmed == "" || trimmed == "null" {
+		return ""
+	}
+
+	var parsed any
+	if err := json.Unmarshal(input, &parsed); err != nil {
+		return trimmed
+	}
+
+	canonical, err := json.Marshal(parsed)
 	if err != nil {
-		return nil, err
+		return trimmed
 	}
 
-	var anthropicResp anthropicResponse
-	err = p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/messages",
-		Body:     anthropicReq,
-	}, &anthropicResp)
-	if err != nil {
-		return nil, err
-	}
-
-	return convertAnthropicResponseToResponses(&anthropicResp, req.Model), nil
+	return string(canonical)
 }
 
-func parseOptionalUnix(ts string) *int64 {
-	ts = strings.TrimSpace(ts)
-	if ts == "" {
-		return nil
+func normalizeAnthropicStopReason(stopReason string) string {
+	switch stopReason {
+	case "tool_use":
+		return "tool_calls"
+	case "end_turn", "stop_sequence":
+		return "stop"
+	case "max_tokens", "model_context_window_exceeded":
+		return "length"
+	default:
+		return stopReason
 	}
-	t, err := time.Parse(time.RFC3339, ts)
-	if err != nil {
-		return nil
-	}
-	u := t.Unix()
-	return &u
-}
-
-func mapAnthropicBatchResponse(resp *anthropicBatchResponse) *core.BatchResponse {
-	if resp == nil {
-		return nil
-	}
-
-	total := resp.RequestCounts.Processing + resp.RequestCounts.Succeeded + resp.RequestCounts.Errored + resp.RequestCounts.Canceled + resp.RequestCounts.Expired
-	failed := resp.RequestCounts.Errored + resp.RequestCounts.Canceled + resp.RequestCounts.Expired
-
-	status := "in_progress"
-	switch resp.ProcessingStatus {
-	case "canceling":
-		status = "cancelling"
-	case "ended":
-		switch {
-		case resp.RequestCounts.Canceled > 0 && resp.RequestCounts.Succeeded == 0 && resp.RequestCounts.Errored == 0:
-			status = "cancelled"
-		case resp.RequestCounts.Errored > 0 && resp.RequestCounts.Succeeded == 0:
-			status = "failed"
-		default:
-			status = "completed"
-		}
-	}
-
-	return &core.BatchResponse{
-		ID:           resp.ID,
-		Object:       "batch",
-		Status:       status,
-		CreatedAt:    parseCreatedAt(resp.CreatedAt),
-		CompletedAt:  parseOptionalUnix(resp.EndedAt),
-		CancellingAt: parseOptionalUnix(resp.CancelInitiatedAt),
-		RequestCounts: core.BatchRequestCounts{
-			Total:     total,
-			Completed: resp.RequestCounts.Succeeded,
-			Failed:    failed,
-		},
-	}
-}
-
-func buildAnthropicBatchCreateRequest(req *core.BatchRequest) (*anthropicBatchCreateRequest, map[string]string, error) {
-	const maxAnthropicBatchRequests = 10000
-
-	if req == nil {
-		return nil, nil, core.NewInvalidRequestError("request is required for anthropic batch processing", nil)
-	}
-	if len(req.Requests) == 0 {
-		return nil, nil, core.NewInvalidRequestError("requests is required for anthropic batch processing", nil)
-	}
-	if len(req.Requests) > maxAnthropicBatchRequests {
-		return nil, nil, core.NewInvalidRequestError("too many requests for anthropic batch processing", nil)
-	}
-
-	out := &anthropicBatchCreateRequest{
-		Requests: make([]anthropicBatchRequest, 0, len(req.Requests)),
-	}
-	endpointByCustomID := make(map[string]string, len(req.Requests))
-	seenCustomIDs := make(map[string]int, len(req.Requests))
-
-	for i, item := range req.Requests {
-		decoded, err := core.DecodeKnownBatchItemRequest(req.Endpoint, item)
-		if err != nil {
-			return nil, nil, core.NewInvalidRequestError(fmt.Sprintf("batch item %d: %s", i, err.Error()), err)
-		}
-
-		params, err := convertDecodedBatchItemToAnthropic(decoded)
-		if err != nil {
-			return nil, nil, prefixAnthropicBatchItemError(i, err)
-		}
-
-		customID := strings.TrimSpace(item.CustomID)
-		if customID == "" {
-			customID = fmt.Sprintf("req-%d", i)
-		}
-		if previousIndex, exists := seenCustomIDs[customID]; exists {
-			return nil, nil, core.NewInvalidRequestError(
-				fmt.Sprintf("batch item %d: duplicate custom_id %q (already used by batch item %d)", i, customID, previousIndex),
-				nil,
-			)
-		}
-		seenCustomIDs[customID] = i
-		out.Requests = append(out.Requests, anthropicBatchRequest{
-			CustomID: customID,
-			Params:   *params,
-		})
-		endpointByCustomID[customID] = decoded.Endpoint
-	}
-
-	return out, endpointByCustomID, nil
-}
-
-func (p *Provider) createBatch(ctx context.Context, req *core.BatchRequest) (*core.BatchResponse, map[string]string, error) {
-	anthropicReq, endpointByCustomID, err := buildAnthropicBatchCreateRequest(req)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	var resp anthropicBatchResponse
-	err = p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/messages/batches",
-		Body:     anthropicReq,
-	}, &resp)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	mapped := mapAnthropicBatchResponse(&resp)
-	if mapped == nil {
-		return nil, nil, core.NewProviderError("anthropic", http.StatusBadGateway, "failed to map anthropic batch response", nil)
-	}
-	mapped.ProviderBatchID = mapped.ID
-	p.setBatchResultEndpoints(mapped.ProviderBatchID, endpointByCustomID)
-	return mapped, cloneBatchResultEndpoints(endpointByCustomID), nil
-}
-
-// CreateBatch creates an Anthropic native message batch.
-func (p *Provider) CreateBatch(ctx context.Context, req *core.BatchRequest) (*core.BatchResponse, error) {
-	mapped, _, err := p.createBatch(ctx, req)
-	return mapped, err
-}
-
-// CreateBatchWithHints creates an Anthropic native message batch and returns
-// persisted per-item endpoint hints for later result shaping.
-func (p *Provider) CreateBatchWithHints(ctx context.Context, req *core.BatchRequest) (*core.BatchResponse, map[string]string, error) {
-	return p.createBatch(ctx, req)
-}
-
-// GetBatch retrieves an Anthropic native message batch.
-func (p *Provider) GetBatch(ctx context.Context, id string) (*core.BatchResponse, error) {
-	var resp anthropicBatchResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodGet,
-		Endpoint: "/messages/batches/" + url.PathEscape(id),
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	mapped := mapAnthropicBatchResponse(&resp)
-	if mapped == nil {
-		return nil, core.NewProviderError("anthropic", http.StatusBadGateway, "failed to map anthropic batch response", nil)
-	}
-	mapped.ProviderBatchID = mapped.ID
-	return mapped, nil
-}
-
-// ListBatches lists Anthropic native message batches.
-func (p *Provider) ListBatches(ctx context.Context, limit int, after string) (*core.BatchListResponse, error) {
-	values := url.Values{}
-	if limit > 0 {
-		values.Set("limit", strconv.Itoa(limit))
-	}
-	// Anthropic uses before_id for reverse-chronological pagination.
-	// Gateway `after` is mapped directly to before_id for provider-native paging.
-	if after != "" {
-		values.Set("before_id", after)
-	}
-	endpoint := "/messages/batches"
-	if encoded := values.Encode(); encoded != "" {
-		endpoint += "?" + encoded
-	}
-
-	var resp anthropicBatchListResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodGet,
-		Endpoint: endpoint,
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-
-	data := make([]core.BatchResponse, 0, len(resp.Data))
-	for _, row := range resp.Data {
-		mapped := mapAnthropicBatchResponse(&row)
-		if mapped == nil {
-			continue
-		}
-		mapped.ProviderBatchID = mapped.ID
-		data = append(data, *mapped)
-	}
-
-	return &core.BatchListResponse{
-		Object:  "list",
-		Data:    data,
-		HasMore: resp.HasMore,
-		FirstID: resp.FirstID,
-		LastID:  resp.LastID,
-	}, nil
-}
-
-// CancelBatch cancels an Anthropic native message batch.
-func (p *Provider) CancelBatch(ctx context.Context, id string) (*core.BatchResponse, error) {
-	var resp anthropicBatchResponse
-	err := p.client.Do(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/messages/batches/" + url.PathEscape(id) + "/cancel",
-	}, &resp)
-	if err != nil {
-		return nil, err
-	}
-	mapped := mapAnthropicBatchResponse(&resp)
-	if mapped == nil {
-		return nil, core.NewProviderError("anthropic", http.StatusBadGateway, "failed to map anthropic batch response", nil)
-	}
-	mapped.ProviderBatchID = mapped.ID
-	return mapped, nil
-}
-
-func (p *Provider) getBatchResults(ctx context.Context, id string, endpointByCustomID map[string]string) (*core.BatchResultsResponse, error) {
-	resp, err := p.client.DoPassthrough(ctx, llmclient.Request{
-		Method:   http.MethodGet,
-		Endpoint: "/messages/batches/" + url.PathEscape(id) + "/results",
-	})
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		body, readErr := io.ReadAll(resp.Body)
-		if readErr != nil {
-			body = []byte("failed to read error response")
-		}
-		return nil, core.ParseProviderError("anthropic", resp.StatusCode, body, nil)
-	}
-
-	scanner := bufio.NewScanner(resp.Body)
-	// Allow larger result lines than Scanner's default 64K.
-	scanner.Buffer(make([]byte, 0, 64*1024), 4*1024*1024)
-	if endpointByCustomID == nil {
-		endpointByCustomID = p.getBatchResultEndpoints(id)
-	} else {
-		endpointByCustomID = cloneBatchResultEndpoints(endpointByCustomID)
-	}
-
-	results := make([]core.BatchResultItem, 0)
-	index := 0
-	for scanner.Scan() {
-		line := bytes.TrimSpace(scanner.Bytes())
-		if len(line) == 0 {
-			continue
-		}
-
-		var row anthropicBatchResultLine
-		if err := json.Unmarshal(line, &row); err != nil {
-			slog.Warn(
-				"failed to decode anthropic batch result line",
-				"error", err,
-				"batch_id", id,
-				"line_index", index,
-				"line_bytes", len(line),
-			)
-			continue
-		}
-		itemEndpoint := "/v1/chat/completions"
-		if endpointByCustomID != nil {
-			if endpoint := strings.TrimSpace(endpointByCustomID[row.CustomID]); endpoint != "" {
-				itemEndpoint = endpoint
-			}
-		}
-
-		item := core.BatchResultItem{
-			Index:    index,
-			CustomID: row.CustomID,
-			URL:      itemEndpoint,
-			Provider: "anthropic",
-		}
-		switch row.Result.Type {
-		case "succeeded":
-			item.StatusCode = http.StatusOK
-			if len(row.Result.Message) > 0 {
-				var anthropicPayload anthropicResponse
-				if err := json.Unmarshal(row.Result.Message, &anthropicPayload); err == nil {
-					switch itemEndpoint {
-					case "/v1/responses":
-						mapped := convertAnthropicResponseToResponses(&anthropicPayload, anthropicPayload.Model)
-						item.Response = mapped
-						item.Model = mapped.Model
-					default:
-						mapped := convertFromAnthropicResponse(&anthropicPayload)
-						item.Response = mapped
-						item.Model = mapped.Model
-					}
-				} else {
-					item.Response = string(row.Result.Message)
-				}
-			}
-		default:
-			item.StatusCode = http.StatusBadRequest
-			errType := row.Result.Type
-			errMsg := "batch item failed"
-			if row.Result.Error != nil {
-				if row.Result.Error.Type != "" {
-					errType = row.Result.Error.Type
-				}
-				if row.Result.Error.Message != "" {
-					errMsg = row.Result.Error.Message
-				}
-			}
-			item.Error = &core.BatchError{
-				Type:    errType,
-				Message: errMsg,
-			}
-		}
-
-		results = append(results, item)
-		index++
-	}
-	if err := scanner.Err(); err != nil {
-		return nil, core.NewProviderError("anthropic", http.StatusBadGateway, "failed to parse anthropic batch results", err)
-	}
-
-	return &core.BatchResultsResponse{
-		Object:  "list",
-		BatchID: id,
-		Data:    results,
-	}, nil
-}
-
-// GetBatchResults retrieves Anthropic native message batch results.
-func (p *Provider) GetBatchResults(ctx context.Context, id string) (*core.BatchResultsResponse, error) {
-	return p.getBatchResults(ctx, id, nil)
-}
-
-// GetBatchResultsWithHints retrieves Anthropic native batch results using
-// persisted per-item endpoint hints instead of transient in-memory state.
-func (p *Provider) GetBatchResultsWithHints(ctx context.Context, id string, endpointByCustomID map[string]string) (*core.BatchResultsResponse, error) {
-	return p.getBatchResults(ctx, id, endpointByCustomID)
-}
-
-// ClearBatchResultHints clears transient per-batch endpoint hints once they
-// have been persisted by the gateway.
-func (p *Provider) ClearBatchResultHints(batchID string) {
-	p.clearBatchResultEndpoints(batchID)
 }
 
 // Embeddings returns an error because Anthropic does not natively support embeddings.
 // Voyage AI (Anthropic's recommended embedding provider) may be added in the future.
 func (p *Provider) Embeddings(_ context.Context, _ *core.EmbeddingRequest) (*core.EmbeddingResponse, error) {
 	return nil, core.NewInvalidRequestError("anthropic does not support embeddings — consider using Voyage AI", nil)
-}
-
-// StreamResponses returns a raw response body for streaming Responses API (caller must close)
-func (p *Provider) StreamResponses(ctx context.Context, req *core.ResponsesRequest) (io.ReadCloser, error) {
-	anthropicReq, err := convertResponsesRequestToAnthropic(req)
-	if err != nil {
-		return nil, err
-	}
-	anthropicReq.Stream = true
-
-	stream, err := p.client.DoStream(ctx, llmclient.Request{
-		Method:   http.MethodPost,
-		Endpoint: "/messages",
-		Body:     anthropicReq,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// Return a reader that converts Anthropic SSE format to Responses API format
-	return newResponsesStreamConverter(stream, req.Model), nil
-}
-
-// responsesStreamConverter wraps an Anthropic stream and converts it to Responses API format
-type responsesStreamConverter struct {
-	reader          *bufio.Reader
-	body            io.ReadCloser
-	model           string
-	responseID      string
-	output          *providers.ResponsesOutputEventState
-	nextOutputIndex int
-	toolCalls       map[int]*providers.ResponsesOutputToolCallState
-	thinkingBlocks  map[int]bool // tracks which content block indices are thinking blocks
-	buffer          streaming.StreamBuffer
-	closed          bool
-	sentDone        bool
-	usage           anthropicUsage
-	hasUsage        bool
-}
-
-func newResponsesStreamConverter(body io.ReadCloser, model string) *responsesStreamConverter {
-	responseID := "resp_" + uuid.New().String()
-	return &responsesStreamConverter{
-		reader:         bufio.NewReader(body),
-		body:           body,
-		model:          model,
-		responseID:     responseID,
-		output:         providers.NewResponsesOutputEventState(responseID),
-		toolCalls:      make(map[int]*providers.ResponsesOutputToolCallState),
-		thinkingBlocks: make(map[int]bool),
-		buffer:         streaming.NewStreamBuffer(1024),
-	}
-}
-
-func (sc *responsesStreamConverter) Read(p []byte) (n int, err error) {
-	if sc.closed {
-		sc.releaseBuffer()
-		return 0, io.EOF
-	}
-
-	// If we have buffered data, return it first
-	if sc.buffer.Len() > 0 {
-		return sc.buffer.Read(p), nil
-	}
-
-	// Read the next SSE event from Anthropic
-	for {
-		line, err := sc.reader.ReadBytes('\n')
-		if err != nil {
-			if err == io.EOF {
-				// Send final done event and [DONE] message
-				if !sc.sentDone {
-					sc.sentDone = true
-					prefix := sc.output.CompleteAssistantOutput(0)
-					responseData := map[string]any{
-						"id":         sc.responseID,
-						"object":     "response",
-						"status":     "completed",
-						"model":      sc.model,
-						"provider":   "anthropic",
-						"created_at": time.Now().Unix(),
-					}
-					// Include merged usage data captured across message_start/message_delta.
-					if sc.hasUsage {
-						responseData["usage"] = anthropicResponsesUsagePayload(&sc.usage)
-					}
-					doneEvent := map[string]any{
-						"type":     "response.completed",
-						"response": responseData,
-					}
-					jsonData, marshalErr := json.Marshal(doneEvent)
-					if marshalErr != nil {
-						slog.Error("failed to marshal response.completed event", "error", marshalErr, "response_id", sc.responseID)
-						sc.closed = true
-						sc.releaseBuffer()
-						_ = sc.body.Close() //nolint:errcheck
-						return 0, io.EOF
-					}
-					sc.buffer.AppendString(prefix)
-					sc.buffer.AppendString("event: response.completed\ndata: ")
-					sc.buffer.AppendBytes(jsonData)
-					sc.buffer.AppendString("\n\ndata: [DONE]\n\n")
-					return sc.buffer.Read(p), nil
-				}
-				sc.closed = true
-				sc.releaseBuffer()
-				_ = sc.body.Close() //nolint:errcheck
-				return 0, io.EOF
-			}
-			return 0, err
-		}
-
-		n, handled, err := consumeAnthropicSSELine(p, line, sc.body, &sc.buffer, sc.convertEvent)
-		if err != nil {
-			sc.closed = true
-			sc.releaseBuffer()
-			return 0, err
-		}
-		if handled {
-			if n == 0 {
-				continue
-			}
-			return n, nil
-		}
-	}
-}
-
-func (sc *responsesStreamConverter) Close() error {
-	if sc.closed {
-		sc.releaseBuffer()
-		return nil
-	}
-	sc.closed = true
-	sc.releaseBuffer()
-	return sc.body.Close()
-}
-
-func (sc *responsesStreamConverter) releaseBuffer() {
-	sc.buffer.Release()
-}
-
-func (sc *responsesStreamConverter) reserveAssistantMessageOutput() {
-	if sc.output.AssistantReserved() {
-		return
-	}
-	sc.output.ReserveAssistant()
-	sc.nextOutputIndex++
-}
-
-func (sc *responsesStreamConverter) newResponsesToolCallState(contentBlock *anthropicContent) *providers.ResponsesOutputToolCallState {
-	callID := providers.ResponsesFunctionCallCallID(contentBlock.ID)
-	state := &providers.ResponsesOutputToolCallState{
-		CallID:      callID,
-		Name:        contentBlock.Name,
-		OutputIndex: sc.nextOutputIndex,
-	}
-	sc.nextOutputIndex++
-
-	initialArguments := extractInitialToolArguments(contentBlock.Input)
-	state.PlaceholderObject = initialArguments == "{}"
-	if initialArguments != "" && !state.PlaceholderObject {
-		_, _ = state.Arguments.WriteString(initialArguments)
-	}
-
-	return state
-}
-
-func (sc *responsesStreamConverter) convertEvent(event *anthropicStreamEvent) string {
-	switch event.Type {
-	case "message_start":
-		if event.Message != nil {
-			if mergeAnthropicUsage(&sc.usage, &event.Message.Usage) {
-				sc.hasUsage = true
-			}
-		}
-		if mergeAnthropicUsage(&sc.usage, event.Usage) {
-			sc.hasUsage = true
-		}
-		// Send response.created event
-		createdEvent := map[string]any{
-			"type": "response.created",
-			"response": map[string]any{
-				"id":         sc.responseID,
-				"object":     "response",
-				"status":     "in_progress",
-				"model":      sc.model,
-				"provider":   "anthropic",
-				"created_at": time.Now().Unix(),
-			},
-		}
-		jsonData, err := json.Marshal(createdEvent)
-		if err != nil {
-			slog.Error("failed to marshal response.created event", "error", err, "response_id", sc.responseID)
-			return ""
-		}
-		return fmt.Sprintf("event: response.created\ndata: %s\n\n", jsonData)
-
-	case "content_block_start":
-		if event.ContentBlock != nil && event.ContentBlock.Type == "thinking" {
-			sc.thinkingBlocks[event.Index] = true
-			return ""
-		}
-		if event.ContentBlock != nil && event.ContentBlock.Type == "tool_use" {
-			if sc.output.AssistantStarted() && !sc.output.AssistantDone() {
-				prefix := sc.output.CompleteAssistantOutput(0)
-				state := sc.newResponsesToolCallState(event.ContentBlock)
-				sc.toolCalls[event.Index] = state
-				return prefix + sc.output.StartToolCall(state, true)
-			}
-			state := sc.newResponsesToolCallState(event.ContentBlock)
-			sc.toolCalls[event.Index] = state
-			return sc.output.StartToolCall(state, true)
-		}
-		return ""
-
-	case "content_block_delta":
-		if event.Delta == nil {
-			return ""
-		}
-
-		switch event.Delta.Type {
-		case "thinking_delta", "signature_delta":
-			// Thinking and signature deltas are part of Anthropic's extended thinking;
-			// the Responses API format does not have a direct equivalent, so skip them.
-			return ""
-		case "text_delta":
-			if event.Delta.Text != "" {
-				sc.reserveAssistantMessageOutput()
-				prefix := sc.output.StartAssistantOutput(0)
-				sc.output.AppendAssistantText(event.Delta.Text)
-				deltaEvent := map[string]any{
-					"type":  "response.output_text.delta",
-					"delta": event.Delta.Text,
-				}
-				jsonData, err := json.Marshal(deltaEvent)
-				if err != nil {
-					slog.Error("failed to marshal content delta event", "error", err, "response_id", sc.responseID)
-					return ""
-				}
-				return prefix + fmt.Sprintf("event: response.output_text.delta\ndata: %s\n\n", jsonData)
-			}
-		case "input_json_delta":
-			if event.Delta.PartialJSON == "" {
-				return ""
-			}
-			state := sc.toolCalls[event.Index]
-			if state == nil {
-				return ""
-			}
-			if state.PlaceholderObject {
-				state.Arguments = strings.Builder{}
-				state.PlaceholderObject = false
-			}
-			_, _ = state.Arguments.WriteString(event.Delta.PartialJSON)
-			return sc.output.WriteEvent("response.function_call_arguments.delta", map[string]any{
-				"type":         "response.function_call_arguments.delta",
-				"item_id":      state.ItemID,
-				"output_index": state.OutputIndex,
-				"delta":        event.Delta.PartialJSON,
-			})
-		}
-		return ""
-
-	case "content_block_stop":
-		state := sc.toolCalls[event.Index]
-		return sc.output.CompleteToolCall(state, true)
-
-	case "message_delta":
-		// Capture usage data for inclusion in response.completed
-		if mergeAnthropicUsage(&sc.usage, event.Usage) {
-			sc.hasUsage = true
-		}
-		if !sc.output.AssistantReserved() && len(sc.toolCalls) == 0 {
-			sc.reserveAssistantMessageOutput()
-		}
-		return ""
-
-	case "message_stop":
-		// Will be handled in Read() when we get EOF
-		return ""
-	}
-
-	return ""
 }
