+++ internal/providers/anthropic/chat.go
@@ -1,0 +1,85 @@
+package anthropic
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"time"
+
+	"gomodel/internal/core"
+	"gomodel/internal/llmclient"
+)
+
+// convertFromAnthropicResponse converts Anthropic response to core.ChatResponse
+func convertFromAnthropicResponse(resp *anthropicResponse) *core.ChatResponse {
+	content := extractTextContent(resp.Content)
+	thinking := extractThinkingContent(resp.Content)
+	toolCalls := extractToolCalls(resp.Content)
+
+	finishReason := normalizeAnthropicStopReason(resp.StopReason)
+	if finishReason == "" {
+		finishReason = "stop"
+	}
+
+	usage := core.Usage{
+		PromptTokens:     resp.Usage.InputTokens,
+		CompletionTokens: resp.Usage.OutputTokens,
+		TotalTokens:      resp.Usage.InputTokens + resp.Usage.OutputTokens,
+	}
+
+	rawUsage := buildAnthropicRawUsage(resp.Usage)
+	if len(rawUsage) > 0 {
+		usage.RawUsage = rawUsage
+	}
+
+	msg := core.ResponseMessage{
+		Role:      "assistant",
+		Content:   content,
+		ToolCalls: toolCalls,
+	}
+
+	// Surface thinking content as reasoning_content (OpenAI-compatible format).
+	if thinking != "" {
+		raw, err := json.Marshal(thinking)
+		if err == nil {
+			msg.ExtraFields = core.UnknownJSONFieldsFromMap(map[string]json.RawMessage{
+				"reasoning_content": raw,
+			})
+		}
+	}
+
+	return &core.ChatResponse{
+		ID:      resp.ID,
+		Object:  "chat.completion",
+		Model:   resp.Model,
+		Created: time.Now().Unix(),
+		Choices: []core.Choice{
+			{
+				Index:        0,
+				Message:      msg,
+				FinishReason: finishReason,
+			},
+		},
+		Usage: usage,
+	}
+}
+
+// ChatCompletion sends a chat completion request to Anthropic
+func (p *Provider) ChatCompletion(ctx context.Context, req *core.ChatRequest) (*core.ChatResponse, error) {
+	anthropicReq, err := convertToAnthropicRequest(req)
+	if err != nil {
+		return nil, err
+	}
+
+	var anthropicResp anthropicResponse
+	err = p.client.Do(ctx, llmclient.Request{
+		Method:   http.MethodPost,
+		Endpoint: "/messages",
+		Body:     anthropicReq,
+	}, &anthropicResp)
+	if err != nil {
+		return nil, err
+	}
+
+	return convertFromAnthropicResponse(&anthropicResp), nil
+}
